trt_requirements.txt
tensorrt-cu12 >= 10.5.0
colored
cuda-python
transformers==4.42.2
opencv-python==4.8.0.74
onnx==1.17.0
onnxruntime==1.19.2
--extra-index-url https://pypi.nvidia.com
nvidia-modelopt[torch,onnx]==0.19.0
onnx-graphsurgeon
polygraphy==0.49.9
sentencepiece
